# Cognitive Execution Layer (CEL)/docs/api/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cognitive Execution Layer (CEL) API",
    "description": "Autonomous Engineering API for AI-powered code development and system management",
    "version": "1.0.0",
    "contact": {
      "name": "CEL Development Team",
      "email": "support@cel-project.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.cel-project.org",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Chat completions endpoint",
        "description": "Generate AI responses for chat interactions",
        "tags": ["AI"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          }
        }
      }
    },
    "/v1/run-tests": {
      "post": {
        "summary": "Run tests on code",
        "description": "Execute tests for the specified code or project",
        "tags": ["Testing"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tests executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lint-code": {
      "post": {
        "summary": "Lint code for issues",
        "description": "Analyze code for style and quality issues",
        "tags": ["Code Quality"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code linting completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["system", "user", "assistant"],
                  "example": "user"
                },
                "content": {
                  "type": "string",
                  "example": "Write a function to calculate fibonacci numbers"
                }
              }
            }
          },
          "model": {
            "type": "string",
            "example": "gpt-4-turbo"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "example": 0.7
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "chatcmpl-123"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string",
                      "example": "assistant"
                    },
                    "content": {
                      "type": "string",
                      "example": "Here's a function to calculate fibonacci numbers..."
                    }
                  }
                },
                "finish_reason": {
                  "type": "string",
                  "example": "stop"
                }
              }
            }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": {
                "type": "number",
                "example": 50
              },
              "completion_tokens": {
                "type": "number",
                "example": 120
              },
              "total_tokens": {
                "type": "number",
                "example": 170
              }
            }
          }
        }
      },
      "TestRequest": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "example": "./src/example.js"
          },
          "testFramework": {
            "type": "string",
            "example": "jest"
          },
          "options": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "TestResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "summary": {
            "type": "string",
            "example": "All tests passed"
          },
          "duration": {
            "type": "number",
            "example": 2.34
          }
        }
      },
      "LintRequest": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "example": "./src/example.js"
          },
          "linter": {
            "type": "string",
            "example": "eslint"
          }
        }
      },
      "LintResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "line": {
                  "type": "number"
                },
                "column": {
                  "type": "number"
                },
                "message": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": ["error", "warning", "info"]
                }
              }
            }
          },
          "summary": {
            "type": "string",
            "example": "2 errors, 5 warnings"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}